+++ cmd/get_imports_test.go
@@ -25,3 +25,36 @@ func SilenceLogs(c cookoo.Context) {
 	p := cookoo.NewParamsWithValues(map[string]interface{}{"quiet": true})
 	BeQuiet(c, p)
 }
+
+func TestGetRepoRootFromPackage(t *testing.T) {
+	urlList := map[string]string{
+		"github.com/Masterminds/VCSTestRepo":                       "github.com/Masterminds/VCSTestRepo",
+		"bitbucket.org/mattfarina/testhgrepo":                      "bitbucket.org/mattfarina/testhgrepo",
+		"launchpad.net/govcstestbzrrepo/trunk":                     "launchpad.net/govcstestbzrrepo/trunk",
+		"launchpad.net/~mattfarina/+junk/mygovcstestbzrrepo":       "launchpad.net/~mattfarina/+junk/mygovcstestbzrrepo",
+		"launchpad.net/~mattfarina/+junk/mygovcstestbzrrepo/trunk": "launchpad.net/~mattfarina/+junk/mygovcstestbzrrepo",
+		"git.launchpad.net/govcstestgitrepo":                       "git.launchpad.net/govcstestgitrepo",
+		"git.launchpad.net/~mattfarina/+git/mygovcstestgitrepo":    "git.launchpad.net/~mattfarina/+git/mygovcstestgitrepo",
+		"farbtastic.googlecode.com/svn/":                           "farbtastic.googlecode.com/svn/",
+		"farbtastic.googlecode.com/svn/trunk":                      "farbtastic.googlecode.com/svn/trunk",
+		"code.google.com/p/farbtastic":                             "code.google.com/p/farbtastic",
+		"code.google.com/p/plotinum":                               "code.google.com/p/plotinum",
+		"example.com/foo/bar.git":                                  "example.com/foo/bar.git",
+		"example.com/foo/bar.svn":                                  "example.com/foo/bar.svn",
+		"example.com/foo/bar/baz.bzr":                              "example.com/foo/bar/baz.bzr",
+		"example.com/foo/bar/baz.hg":                               "example.com/foo/bar/baz.hg",
+		"gopkg.in/foo.v1":                                          "gopkg.in/foo.v1",
+		"gopkg.in/foo.v1/bar":                                      "gopkg.in/foo.v1",
+		"gopkg.in/baz/foo.v1":                                      "gopkg.in/baz/foo.v1",
+		"gopkg.in/baz/foo.v1/bar":                                  "gopkg.in/baz/foo.v1",
+		"golang.org/x/net":                                         "golang.org/x/net",
+		"golang.org/x/net/context":                                 "golang.org/x/net",
+	}
+
+	for u, c := range urlList {
+		repo := getRepoRootFromPackage(u)
+		if repo != c {
+			t.Errorf("getRepoRootFromPackage expected %s but got %s", c, repo)
+		}
+	}
+}