+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Release 0.5.1 (xxxx-xx-xx)
+
+- Fixes #58: Guess command not working.
+- Fixes #56: Unable to use glide get on golang.org/x/[name]/[subpackage]
+
 # Release 0.5.0 (2015-08-19)
 
 **Glide .5 is a major update breaking some backwards compatability with